exitコマンドも綺麗かどうかは別として終わった

+++ execute.h
@@ -6,7 +6,7 @@
 /*   By: toshi <toshi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/30 18:29:28 by tofujiwa          #+#    #+#             */
-/*   Updated: 2024/04/08 20:31:56 by toshi            ###   ########.fr       */
+/*   Updated: 2024/04/08 20:55:04 by toshi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -239,7 +239,7 @@ char *expand_path(char *oldpath, char *newpath);
 //~~~~~~~~
 
 //~~~~exit start~~~~
-t_bool	is_longlong_over(const char *str, long long *ans_num);
+t_bool	is_longlong_over(const char *str, int *ans_num);
 int do_exit(char **cmd_args, t_manager *manager);
 //~~~~~~~~
